+++ chapter01/problem_10/main.cpp
@@ -13,7 +13,7 @@ std::ostream& operator<<(std::ostream& os, const container& c) {
 
 std::vector<std::string> gray_code(size_t num) {
   std::vector<std::string> codes;
-  for (size_t i = 0; i < (1 << num); ++i) {
+  for (size_t i = 0; i < (static_cast<size_t>(1) << num); ++i) {
     auto value = (i ^ (i >> 1));
     codes.emplace_back(std::bitset<32>(value).to_string().substr(32 - num));
   }
